Set "copy to output directory" option for mystem.exe
